Comments, UX and cleanup
